requirements.txt
dj-database-url==0.5.0
Django==2.1.5
django-filter==2.1.0
gunicorn==19.9.0
pkg-resources==0.0.0
psycopg2==2.7.4
python-decouple==3.1
pytz==2018.9
whitenoise==4.1.2
